fix(tests,repo): land the gates the environment deferral had hidden
Retrying the three deferred gates in a supported environment surfaced three
real defects, all fixed here rather than waived.

Release-surface pins: the two tests/roster release-synchronization tests
hard-pin the previous core and agentbundle versions, so the bump left them red.
Both now name the shipped versions.

Pack test boundary: the seed contract file asserted against this checkout's
root AGENTS.md, root docs/architecture/overview.md, and an adopter-facing
how-to, which pack-tests-stay-in-pack rejects. The repository-level half moves
to tests/roster/test_repository_context_root_guidance.py; the seed half stays
pack-owned and anchors at its own pack root. tests/conformance/ cannot host it
— its portability lint rejects any reach into docs/.

Suite runners: contract-acquisition and architect-design were discovered
rather than declared, which every-suite-dir-has-a-runner treats as unrun. Both
are now named in the Makefile test target.

Engine-Change-RFC: n/a — no change under packages/agentbundle/agentbundle/;
this commit touches tests, the Makefile, spec notes, and generated highlights.
