+++ Makefile
@@ -398,6 +398,7 @@ test-unleased: lint-editable-install
 	$(PYTHON) -m pytest packs/core/tests/skills/author-brief/ -q
 	$(PYTHON) -m pytest packs/core/tests/skills/bug-fix/ -q
 	$(PYTHON) -m pytest packs/core/tests/skills/capture-work/ -q
+	$(PYTHON) -m pytest packs/core/tests/skills/contract-acquisition/ -q
 	$(PYTHON) -m pytest packs/core/tests/skills/new-spec/ -q
 	$(PYTHON) -m pytest packs/core/tests/skills/project-knowledge/ -q
 	$(PYTHON) -m pytest packs/core/tests/skills/receive-brief/ -q
@@ -409,6 +410,7 @@ test-unleased: lint-editable-install
 	$(PYTHON) -m pytest packs/product-documentation/tests/ -q
 	$(PYTHON) -m pytest packs/architect/tests/pack/ -q
 	$(PYTHON) -m pytest packs/architect/tests/skills/architect-assess/ -q
+	$(PYTHON) -m pytest packs/architect/tests/skills/architect-design/ -q
 	$(PYTHON) -m pytest packs/architect/tests/skills/architect-review/ -q
 	$(PYTHON) -m pytest packs/credential-brokers/tests/pack/ -q
 	$(PYTHON) -c "import httpx"
