+++ tests/roster/test_okf_catalogue_discovery.py
@@ -37,7 +37,7 @@ def _assert_show_schema(response: dict[str, object]) -> None:
 
 def test_release_metadata_moves_together_for_okf_catalogue_discovery() -> None:
     """Public show JSON changes must move repository release surfaces together."""
-    expected = "0.39.0"
+    expected = "0.39.1"
     pyproject = tomllib.loads(
         (PACKAGE_ROOT / "pyproject.toml").read_text(encoding="utf-8")
     )
